requirements.txt
anyio==4.0.0
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
libretranslatepy==2.1.1
lxml==4.9.3
Markdown==3.4.4
Pillow==10.0.1
pyTelegramBotAPI==4.14.0
python-telegram-bot==20.6
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
translate==3.6.1
tzdata==2023.3
urllib3==2.0.5
